requirements.txt
fastapi
uvicorn
sentence-transformers
faiss-cpu
torch
numpy






